Almost done
I tried implementing functions to set all empty tiles to a space background, I chsnged the designs of the missile and players sprites for the new background, I tried to implement dificulty setting s by changeing the interval speed for the movemennt of missiels which failed. I tried a lot of diffeerent options for these to work, and I'm very close to succeeding.